#434 merge composite models
