+++ docs/source/models/lead_acid/composite.rst
@@ -3,6 +3,3 @@ Lead Acid Composite Models
 
 .. autoclass:: pybamm.lead_acid.Composite
     :members:
-
-.. autoclass:: pybamm.lead_acid.CompositeCapacitance
-    :members:
