+++ pybamm/models/lead_acid/__init__.py
@@ -4,4 +4,3 @@
 from .loqs import LOQS
 from .composite import Composite
 from .newman_tiedemann import NewmanTiedemann
-from .composite_capacitance import CompositeCapacitance
