+++ tests/integration/test_models/test_lead_acid/test_composite.py
@@ -3,6 +3,7 @@
 #
 import pybamm
 import tests
+from pybamm.solvers.scikits_ode_solver import scikits_odes_spec
 
 import unittest
 import numpy as np
@@ -16,6 +17,35 @@ class TestLeadAcidComposite(unittest.TestCase):
 
     def test_optimisations(self):
         model = pybamm.lead_acid.Composite()
+        optimtest = tests.OptimisationsTest(model)
+
+        original = optimtest.evaluate_model()
+        simplified = optimtest.evaluate_model(simplify=True)
+        using_known_evals = optimtest.evaluate_model(use_known_evals=True)
+        simp_and_known = optimtest.evaluate_model(simplify=True, use_known_evals=True)
+        np.testing.assert_array_almost_equal(original, simplified)
+        np.testing.assert_array_almost_equal(original, using_known_evals)
+        np.testing.assert_array_almost_equal(original, simp_and_known)
+
+
+class TestLeadAcidCompositeCapacitance(unittest.TestCase):
+    def test_basic_processing_differential(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.Composite(options)
+        modeltest = tests.StandardModelTest(model)
+        modeltest.test_all()
+
+    @unittest.skipIf(scikits_odes_spec is None, "scikits.odes not installed")
+    def test_basic_processing_algebraic(self):
+        options = {"capacitance": "algebraic"}
+        model = pybamm.lead_acid.Composite(options)
+        modeltest = tests.StandardModelTest(model)
+
+        modeltest.test_all()
+
+    def test_optimisations(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.Composite(options)
         optimtest = tests.OptimisationsTest(model)
 
         original = optimtest.evaluate_model()
