+++ tests/unit/test_models/test_lead_acid/test_composite.py
@@ -11,6 +11,26 @@ class TestLeadAcidComposite(unittest.TestCase):
         model.check_well_posedness()
 
 
+class TestLeadAcidCompositeCapacitance(unittest.TestCase):
+    def test_well_posed_differential(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.Composite(options)
+        model.check_well_posedness()
+
+    def test_well_posed_algebraic(self):
+        options = {"capacitance": "algebraic"}
+        model = pybamm.lead_acid.Composite(options)
+        model.check_well_posedness()
+
+    def test_default_solver(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.Composite(options)
+        self.assertIsInstance(model.default_solver, pybamm.ScikitsOdeSolver)
+        options = {"capacitance": "algebraic"}
+        model = pybamm.lead_acid.Composite(options)
+        self.assertIsInstance(model.default_solver, pybamm.ScikitsDaeSolver)
+
+
 if __name__ == "__main__":
     print("Add -v for more debug output")
     import sys
